requirements.txt
jupyterlab
ipykernel
ipywidgets
openai
python-dotenv
streamlit
pandas
pydantic
plotly